refactor request body parsing
